ajout de fonctions

+++ Libft/libft.h
@@ -6,7 +6,7 @@
 /*   By: hugur <marvin@42lausanne.ch>               +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/24 18:37:58 by hugur             #+#    #+#             */
-/*   Updated: 2022/10/31 22:17:35 by hugur            ###   ########.fr       */
+/*   Updated: 2022/11/01 15:14:45 by hugur            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,6 +15,7 @@
 
 #include <stdio.h>
 #include <ctype.h>
+#include <stdlib.h>
 
 int	ft_isdigit(int c);
 int ft_isalpha(int c);
@@ -28,8 +29,8 @@ int ft_strlcpy(char *dst, char *src, int dstsize);
 int ft_atoi(const char *str);
 void *ft_memset(void *b, int c, size_t len);
 char *ft_strchr(const char *s, int c);
-int ft_strncmp(const char *s1, const char *s2, int n);
-void ft_bzero(void *s, int n);
+int ft_strncmp(const char *s1, const char *s2, size_t n);
+void ft_bzero(void *s, size_t n);
 void *ft_memcpy(void *dst, const void *src, int n);
 void	*ft_memmove(void *dest, const void *src, size_t count);
 size_t  ft_strlcat(char *dst, const char *src, size_t size);
